+++ alinkjs/packages/assetlink-pwa/vue.config.js
@@ -1,3 +1,5 @@
+const fs = require("fs");
+const https = require("https");
 const { defineConfig } = require("@vue/cli-service");
 const { ModuleFederationPlugin } = require("webpack").container;
 
@@ -22,12 +24,18 @@ const createDevServerConfig = () => {
         ws: false,
         target: DEV_PROXY_TARGET,
         context: () => true,
-        secure: targetUrl.protocol === "https",
+        secure: targetUrl.protocol === "https:",
         changeOrigin: true,
         bypass(req) {
           if (req.path.indexOf("/alink/sidecar") === 0) {
             console.log(
               `'${req.path}' is an alink sidecar url - passing to proxy...`
+            );
+            return null;
+          }
+          if (req.path.indexOf("/alink/plugins") === 0) {
+            console.log(
+              `'${req.path}' is an alink plugins url - passing to proxy...`
             );
             return null;
           }
@@ -64,7 +72,14 @@ const createDevServerConfig = () => {
     },
   };
 
-  if (targetUrl.protocol === "https") {
+  if (targetUrl.protocol === "https:") {
+    const devRootCA = `${__dirname}/../../../devcerts/rootCA.pem`;
+
+    https.globalAgent.options.ca = https.globalAgent.options.ca || [];
+    https.globalAgent.options.ca.push(fs.readFileSync(devRootCA));
+
+    serverConfig.proxy["^/"].agent = https.globalAgent;
+
     Object.assign(serverConfig, {
       // Deprecated `https` config still needed to cause links printed to console to have correct protocol
       // https://github.com/symfony/webpack-encore/issues/1064
@@ -72,9 +87,9 @@ const createDevServerConfig = () => {
       server: {
         type: "https",
         options: {
-          ca: `${__dirname}/../devcerts/rootCA.pem`,
-          key: `${__dirname}/../devcerts/${devHost}/privkey.pem`,
-          cert: `${__dirname}/../devcerts/${devHost}/fullchain.pem`,
+          ca: devRootCA,
+          key: `${__dirname}/../../../devcerts/${devHost}/privkey.pem`,
+          cert: `${__dirname}/../../../devcerts/${devHost}/fullchain.pem`,
         },
       },
       host: devHost,