+++ CHANGELOG.md
@@ -18,6 +18,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Re-add service-worker.js which got lost in the migration work
 - Fixed logic to connect/disconnect Orbit.js JSON:API source queries based on online/offline status
 - Fix reloading plugins so the `skipCache` parameter also skips the service worker cache
+- Fix handling of dev certs when proxying https farmOS instances for development
 
 ## [0.5.2] - 2022-08-16