+++ CHANGELOG.md
@@ -7,6 +7,22 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.5.3] - 2022-08-17
+
+### Added
+
+- Implement service worker refresh UI to cleanly update the PWA
+- Add reload plugin button to the manage plugins page
+- Add a more specific warning message for asset proximity search when location permissions are denied
+
+### Fixed
+
+- Re-add service-worker.js which got lost in the migration work
+- Fixed logic to connect/disconnect Orbit.js JSON:API source queries based on online/offline status
+- Fix reloading plugins so the `skipCache` parameter also skips the service worker cache
+- Fix handling of dev certs when proxying https farmOS instances for development
+- Actually show messages that are pushed into `assetLink.vm.messages`
+
 ## [0.5.2] - 2022-08-16
 
 ### Added